+++ Main/GameLogicArchitecture/Commons/StateCloner/IStateCloner.cs
@@ -4,6 +4,6 @@ namespace VMFramework.GameLogicArchitecture
 {
     public interface IStateCloner
     {
-        public void CloneFrom(IStateCloner cloner, StateCloneHint hint);
+        public void CloneFrom(IStateCloner cloner, StateCloneContext context);
     }
 }