+++ Main/GameLogicArchitecture/Commons/StateCloner/StateCloneTags.cs
@@ -1,0 +1,7 @@
+namespace VMFramework.GameLogicArchitecture
+{
+    public static class StateCloneTags
+    {
+        public static StateCloneTag OwnerStateIncluded { get; } = StateCloneTag.Create();
+    }
+}